Module convex_hull_simulation::algorithms

Contains the implementation of the algorithms used in the simulation.

Currently, the simulation supports two algorithms:

Jarvis March

Kirkpatrick Seidel

Furthermore contains algorithm relevant functions.

Structs

Algorithm

Bevy resource that contains the current algorithm being used

ConvexHull

Bevy component representing points part of convex hull

DrawingHistory

Bevy resource that contains all the point history, so that they can be animated later. Support all primitives under LineType.

Gizmo

Bevy component that represents temporary points

Enums

AlgorithmType

Enum representing the implemented algorithms

HullType 🔒

Represents the type of hull being calculated. Used for drawing purposes only in kirk_patrick_seidel.

LineType

Enum representing the different types of draw calls the simulation can make

Orientation 🔒

Represent the orientation between three points (consecutive)

Functions

bridge 🔒

Returns the two points that forms the bridge of the points

connect 🔒

Returns the points that form the convex hull

jarvis_march

Implementation of the Jarvis March algorithm (Gift-Wrapping algorithm)

kirk_patrick_seidel

Implementation of the Kirkpatrick Seidel Algorithm

median_of_medians

Returns the Median of medians of the input list

orientation 🔒

Finds the orientation of three points and returns Orientation

upper_hull 🔒

Returns the upper hull of input points