Function convex_hull_simulation::algorithms::bridge

fn bridge(points: &[Vec2], median: f32) -> (Vec2, Vec2)

Returns the two points that forms the bridge of the points

Pseudocode

Function BRIDGE (S, a)
0. CANDIDATES := ∅
1. If |S| = 2 then return ((i, j)), where S = {p_i, p_j} and x(p_i) < x(p_j).
2. Choose ⌊|S|/2⌋ disjoint sets of size 2 from S.
    If a point of S remains, then insert it into CANDIDATES.
    Arrange each subset to be an ordered pair (p_i, p_j), such that x(p_i) < x(p_j).
    Let PAIRS be the set of these ordered pairs.
3. Determine the slopes of the straight lines defined by the pairs.
    In case the slope does not exist for some pair, apply Lemma 3.1, i.e.
    "For all (p_i, p_j) in PAIRS do
        if x(p_i) = x(p_j) then delete (p_i, p_j) from PAIRS
        if y(p_i) > y(p_j) then insert p_i into CANDIDATES
        else insert p_j into CANDIDATES
        else let k(p_i, p_j) := (y(p_j) - y(p_i)) / (x(p_j) - x(p_i))"
4. Determine K, the median of {k(p_i, p_j) | (p_i, p_j) ∈ PAIRS}.
5. Let SMALL := {(p_i, p_j) ∈ PAIRS | k(p_i, p_j) < K}.
    Let EQUAL := {(p_i, p_j) ∈ PAIRS | k(p_i, p_j) = K}.
    Let LARGE := {(p_i, p_j) ∈ PAIRS | k(p_i, p_j) > K}.
6. Find the set of points of S which lie on the supporting line h with slope K, i.e.:
    Let MAX be the set of points p ∈ S, s.t. y(p) - K * x(p) is maximum.
    Let p_k be the point in MAX with minimum x-coordinate.
    Let p_m be the point in MAX with maximum x-coordinate.
7. Determine if h contains the bridge, i.e.
    if x(p_k) ≤ a and x(p_m) > a then return((k, m)).
8. h contains only points to the left of or on L:
    if x(p_m) ≤ a then
        for all (p_i, p_j) ∈ LARGE ∪ EQUAL insert p_j into CANDIDATES.
        for all (p_i, p_j) ∈ SMALL insert p_i and p_j into CANDIDATES.
9. h contains only points to the right of L:
    if x(p_k) > a then
        for all (p_i, p_j) ∈ SMALL ∪ EQUAL insert p_i into CANDIDATES.
        for all (p_i, p_j) ∈ LARGE insert p_i and p_j into CANDIDATES.
10. return(BRIDGE (CANDIDATES, a)).