Function convex_hull_simulation::algorithms::median_of_medians

pub fn median_of_medians<T>(nums: &[T]) -> T
where
    T: Clone + Copy + PartialOrd,

Returns the Median of medians of the input list

Pseudocode

Function MEDIAN-OF-MEDIANS(S)
1. If |S| ≤ 5 then return the median of S.
2. Partition S into ⌈|S|/5⌉ disjoint sets of size 5.
3. Let M be the set of medians of the sets.
4. return MEDIAN-OF-MEDIANS(M).

Analysis

The algorithm is a divide-&-conquer algorithm. The key step is the computation of the median of medians which is based on the partitioning of the input set into disjoint sets of size 5. This step can be done in O(n) time. We also know that step 4 can be done in O(n) time by the linear time median finding algorithm. Hence, steps 2-4 can be done in O(n) time. For the purposes of analyzing algorithm MEDIAN-OF-MEDIANS(S), let us assume the input set consists of n elements. Let $T(n)$ denote the worst-case time complexity of the algorithm. Suppose the two occurrences of $O(n)$ in the above recurrence are at most $cn$, where $c$ is a suitably large constant. We will show by induction on $n$ that $T(n) \leq cn \log(n)$ for all $n \geq 5$. For the base case where $n = 5$, $T(n) \leq cn \leq cn\log(5) = cn \log(n)$. For the inductive case, $T(n) \leq cn + T(\frac{n}{5})$. By the inductive hypothesis, $T(\frac{n}{5}) \leq c \frac{n}{5} \log(\frac{n}{5})$. Therefore, $T(n) \leq cn + c \frac{n}{5} \log(\frac{n}{5}) = cn \log(n)$, completing the induction.