Module convex_hull_simulation::distributions

Contains the different types of distributions that can be used to place the points

The distributions are:

Fibonacci

Random

Structs

Distribution

A resource that stores the current distribution type

Enums

DistributionType

The different types of distributions that can be used to place the points

Constants

GOLDEN_ANGLE 🔒

Stores the value for Golden Angle. Used by fibonacci_circle function to generate points in a fibonacci spiral.

Functions

bounded_random

Generates a random point within a circle with a radius of num_shapes * 25.0

bounded_random_square

Generates points inside a square

circle_points

Generates points on the perimeter of circle

fibonacci_circle

Generate a point in the fibonacci spiral at the given index